requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/City-of-Helsinki/django-helusers.git#egg=django-helusers
defusedxml==0.4.1         # via python3-openid
django-allauth==0.27.0
Django==1.10              # via django-allauth
djangorestframework==3.4.6
markdown==2.6.6
oauthlib==1.1.2           # via requests-oauthlib
py==1.4.31                # via pytest
pytest-django==3.0.0
pytest==3.0.1
python3-openid==3.0.10    # via django-allauth
requests-cache==0.4.12
requests-oauthlib==0.6.2  # via django-allauth
requests==2.11.1
